Fix to solve issue #627
This change is to ensure that the permissions set for whether a page is viewable by the unauthenticated public is considered in the logic when displaying the page. Currently, only the community's permissions are considered in the check, and this change will use a variable already being set to ensure that both permissions are considered. Closes #627