Removed erroneous key being used int he `draft_events` table which refers to a non-existent column.
Added missing check to ensure the Permissions Mask sidebar doesn't show up unless appropriate on the admin side, as well as an additional check for the active status of the user_access record, rather than assuming it's active in the User->get() method.

Also added an isset() for a value check in the events module to stop errors from being thrown many times on every page load if "last_visited" is null.