Baseline Clerkship functionality should work now, with most logic being brought over from the various changes to how things work at Queen's from the last 3-4 years. Additionally, the ability to manage the clerkship rotation categories is now within the admin side of the module to allow building of Streams/Blocks/Rotations as required by the scheduling component of individual schools.
Additional polish will be required, and some interfaces may need to be updated to mesh properly with twitter bootstrap, but this should make clerkship work "out of the box" as opposed to the past when data had to be added directly to the database to even allow the use of the clerkship module at all. Further testing will also be done, but all functionality was working when cursory testing was done on my end.