Refs #410, if user notification is selected while editing a user profile, the email is now sent.
In settings.inc.php, there is now a new variable for the email body text when sending a user a notification that their profile has been updated (separate from the new user notification text). The rich text editor is disabled for the email body text when editing a user as it is when adding one, since CKEditor is html-based and doesn't respect whitespace. This way the formatting from the text in settings.inc.php is preserved without adding html formatting.