Copyright policy acceptance. Check for the user's acceptance of the school's copyright policy upon first login and upon community file and photo uploads.
Copyright information is stored in the active template language file and the user's acknowledgement date is stored in the `user_data`.`copyright` field.

Upon login: if, in the .lang.php file,  there is a valid 'copyright-firstlogin' notice and a 'copyright-version' date that is later than the that in `user_data`.`copyright`
the user is directed to the firstlogin page to agree with the copyright policy. The time of acceptance is stored in the `user_data`.`copyright` field.

Upon uploading Community photos or files: if there is a valid 'copyright-uploads' notice in the language file the user has to accept the terms before they can proceed with the upload.