Ref #275, Adds an ICS download/subscribe option for individual courses.
An ICS file for an individual course can now be downloaded or subscribed to from the dashboard calendar. Currently faculty/staff don't see the calendar. At UCLA we changed that so faculty can see a calendar, which options to change the course shown on it. If that feature is desired also, I can bring it over. Currently on the github instance they can't get to the interface I added. The button on "My Teaching Events" is the ICS for their own teaching events still.