Added the ability to upload multiple files for an assignment.
When creating an assignment, there is an option to allow the learners to upload multiple files, and if this is selected the admin can set a maximum number of files. Because of this, submissions with multiple files are downloaded as .zip archives, and when downloading all submissions at once the individual student submissions are put into folders within the archive. The commenting system had to be revised as well, and now comments are only attached to the assignment and the learner's proxy_id, not the individual afile_id's.