Refs #303, fixes success message for activating learners in Manage Cohorts.
When activating/deactivating/deleting learners from a group, the repeated message "Successfully activated the selected group member" is replaced by "Successfully activated <name of learner>".

On the Edit Groups interface, the "Delete/Activate" button has been replaced by a green "Activate" button and a red "Delete/Deactivate" button. When the user clicks on one of these it passes their choice via POST (not GET since the form is already submitting via POST) so that on the confirmation page there is either a green "Activate" button or a pair of "Deactivate" and "Delete" buttons, but not both.