Ref #227, Display content access statistics.
This commit brings a way to show the access statistics in the interface for items: event viewing, event content download stats, grade book views, document sharing views/downloads. It also cleans up the history_log() done in the learning events section, with an entry now created when the event is created, either through add, copy, or draft. In the past the changes weren't logged until a second person worked on the event.